cosine similarity for city ranks
basic prototype with JS. The scores do not change a lot when they are
high - cosine flattens significantly as it gets to pi/2 and sometimes
the values get 'stuck' around there until you mess around with the
sliders. Perhaps measuring the tan of the angle between the two vectors
and averaging with cosine similarity would be better? We'd get an
average between dy/dx | pi/2 = 0 (for cosine) and dy/dx |pi/2 = inf
(think kinda like line through a gini coefficient curve). Or perhaps we
could figure out some smoothing